Project groupId and java package changed. Preparation for deployment to maven repo.
