--- aspect/src/main/java/com/github/wpik/extramdc/annotation/ExtraMdc.java
@@ -1,11 +1,0 @@
-package com.github.wpik.extramdc.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.METHOD)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface ExtraMdc {
-}
