--- aspect/src/main/java/com/github/wpik/extramdc/annotation/MdcFields.java
@@ -1,12 +1,0 @@
-package com.github.wpik.extramdc.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.PARAMETER)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface MdcFields {
-    MdcField[] value();
-}
