--- aspect/src/main/java/com/github/wpik/extramdc/expression/ExpressionResolver.java
@@ -1,5 +1,0 @@
-package com.github.wpik.extramdc.expression;
-
-public interface ExpressionResolver {
-    String resolve(String expression, Object object);
-}
