+++ aspect/src/main/java/io/github/wpik/extramdc/annotation/ExtraMdc.java
@@ -1,0 +1,11 @@
+package io.github.wpik.extramdc.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface ExtraMdc {
+}
