+++ aspect/src/main/java/io/github/wpik/extramdc/annotation/MdcField.java
@@ -1,0 +1,18 @@
+package io.github.wpik.extramdc.annotation;
+
+import org.springframework.core.annotation.AliasFor;
+
+import java.lang.annotation.*;
+
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+@Repeatable(MdcFields.class)
+public @interface MdcField {
+    @AliasFor("name")
+    String value() default "";
+
+    @AliasFor("value")
+    String name() default "";
+
+    String expression() default "";
+}
