+++ aspect/src/main/java/io/github/wpik/extramdc/annotation/MdcFields.java
@@ -1,0 +1,12 @@
+package io.github.wpik.extramdc.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface MdcFields {
+    MdcField[] value();
+}
