+++ aspect/src/main/java/io/github/wpik/extramdc/expression/ExpressionResolver.java
@@ -1,0 +1,5 @@
+package io.github.wpik.extramdc.expression;
+
+public interface ExpressionResolver {
+    String resolve(String expression, Object object);
+}
