+++ example/pom.xml
@@ -4,14 +4,14 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 
     <parent>
-        <groupId>com.github.wpik</groupId>
+        <groupId>io.github.wpik</groupId>
         <artifactId>extra-mdc</artifactId>
         <version>1.0.0-SNAPSHOT</version>
     </parent>
 
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.github.wpik.extra-mdc</groupId>
+    <groupId>io.github.wpik.extra-mdc</groupId>
     <artifactId>example</artifactId>
 
     <dependencies>
@@ -20,7 +20,7 @@
             <artifactId>spring-boot-starter-web</artifactId>
         </dependency>
         <dependency>
-            <groupId>com.github.wpik.extra-mdc</groupId>
+            <groupId>io.github.wpik.extra-mdc</groupId>
             <artifactId>spring-boot-starter</artifactId>
         </dependency>
         <dependency>
