--- example/src/main/java/com/github/wpik/extramdc/example/Application.java
@@ -1,13 +1,0 @@
-package com.github.wpik.extramdc.example;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.EnableAspectJAutoProxy;
-
-@EnableAspectJAutoProxy
-@SpringBootApplication
-public class Application {
-    public static void main(String[] args) {
-        SpringApplication.run(Application.class, args);
-    }
-}
