+++ example/src/main/resources/application.yml
@@ -1,5 +1,5 @@
 logging:
   level:
-    com.github.wpik: debug
+    io.github.wpik: debug
   pattern:
     level: '%5p [%X]'
