+++ pom.xml
@@ -2,7 +2,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 
     <modelVersion>4.0.0</modelVersion>
-    <groupId>com.github.wpik</groupId>
+    <groupId>io.github.wpik</groupId>
     <artifactId>extra-mdc</artifactId>
     <packaging>pom</packaging>
     <name>extra-mdc</name>
@@ -51,12 +51,12 @@
                 <scope>import</scope>
             </dependency>
             <dependency>
-                <groupId>com.github.wpik.extra-mdc</groupId>
+                <groupId>io.github.wpik.extra-mdc</groupId>
                 <artifactId>aspect</artifactId>
                 <version>1.0.0-SNAPSHOT</version>
             </dependency>
             <dependency>
-                <groupId>com.github.wpik.extra-mdc</groupId>
+                <groupId>io.github.wpik.extra-mdc</groupId>
                 <artifactId>spring-boot-starter</artifactId>
                 <version>1.0.0-SNAPSHOT</version>
             </dependency>
