+++ spring-boot-starter/pom.xml
@@ -5,18 +5,18 @@
 
     <parent>
         <artifactId>extra-mdc</artifactId>
-        <groupId>com.github.wpik</groupId>
+        <groupId>io.github.wpik</groupId>
         <version>1.0.0-SNAPSHOT</version>
     </parent>
 
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.github.wpik.extra-mdc</groupId>
+    <groupId>io.github.wpik.extra-mdc</groupId>
     <artifactId>spring-boot-starter</artifactId>
 
     <dependencies>
         <dependency>
-            <groupId>com.github.wpik.extra-mdc</groupId>
+            <groupId>io.github.wpik.extra-mdc</groupId>
             <artifactId>aspect</artifactId>
         </dependency>
