+++ spring-boot-starter/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-  com.github.wpik.extramdc.spring.ExtraMdcAutoConfiguration
+  io.github.wpik.extramdc.spring.ExtraMdcAutoConfiguration
